requirements/project.txt
Django==1.3.1
south
django-registration
django-avatar
django-profiles
sorl-thumbnail==3.2.5
django-tagging # for gallery
django-simple-captcha
django-netauth
django-form-utils # admin image
django-crispy-forms
git+git://github.com/nicolaslara/django-genericadmin.git#egg=django-genericadmin
git+git://github.com/dfalk/django-inlines.git#egg=django-inlines
BeautifulSoup
django-pagination
git+git://github.com/dfalk/django-cart.git#egg=django-cart
django-mptt
markdown
mezzanine
git+git://github.com/dfalk/mezzanine-wiki.git@dev#egg=mezzanine-wiki
diff-match-patch
django-compressor
django-dbtemplates


### deprecated

#basic.inlines # django-inlines now
#django-seo
#django-haystack
#django-messages
#-e git+git@github.com/arneb/django-messages.git#egg=django-messages
#djangobb_forum
#git+git://github.com/slav0nic/DjangoBB.git#egg=djangobb
#simplepages
#git+git://github.com/alekam/django-simplepages.git#egg=django-simplepages
#mysql-python
